initial data structure changes for 3dcnn, conv3d, pooling3d (#3378)

+++ src/layer/arm/convolution_3x3_int8.h
@@ -18,7 +18,7 @@
 
 static void conv3x3s1_winograd23_transform_kernel_int8_neon(const Mat& kernel, std::vector<Mat>& kernel_tm2, int inch, int outch, const Option& opt)
 {
-    Mat kernel_tm(4 * 4, inch, outch, 2ul);
+    Mat kernel_tm(4 * 4, inch, outch, (size_t)2u);
 
     // G
     const short ktm[4][3] = {
@@ -65,7 +65,7 @@ static void conv3x3s1_winograd23_transform_kernel_int8_neon(const Mat& kernel, s
 
     for (int r = 0; r < 4; r++)
     {
-        Mat kernel_tm_test(4 * 8, inch, outch / 8 + (outch % 8) / 4 + outch % 4, 2u);
+        Mat kernel_tm_test(4 * 8, inch, outch / 8 + (outch % 8) / 4 + outch % 4, (size_t)2u);
 
         int p = 0;
         for (; p + 7 < outch; p += 8)
@@ -1061,7 +1061,7 @@ static void conv3x3s1_winograd23_int8_neon(const Mat& bottom_blob, Mat& top_blob
 
 static void conv3x3s1_winograd43_transform_kernel_int8_neon(const Mat& kernel, std::vector<Mat>& kernel_tm2, int inch, int outch, const Option& opt)
 {
-    Mat kernel_tm(6 * 6, inch, outch, 2ul);
+    Mat kernel_tm(6 * 6, inch, outch, (size_t)2u);
 
     // G
     // const float ktm[6][3] = {
@@ -1118,7 +1118,7 @@ static void conv3x3s1_winograd43_transform_kernel_int8_neon(const Mat& kernel, s
 
     for (int r = 0; r < 9; r++)
     {
-        Mat kernel_tm_test(4 * 8, inch, outch / 8 + (outch % 8) / 4 + outch % 4, 2u);
+        Mat kernel_tm_test(4 * 8, inch, outch / 8 + (outch % 8) / 4 + outch % 4, (size_t)2u);
 
         int p = 0;
         for (; p + 7 < outch; p += 8)
