+++ tests/test_mat_pixel.cpp
@@ -58,7 +58,7 @@ static int test_mat_pixel_gray(int w, int h)
     for (int i = 0; i < 1; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels(a, pixel_type_from[i], w, h);
-        ncnn::Mat b(w, h, 1u, 1);
+        ncnn::Mat b(w, h, (size_t)1u, 1);
         m.to_pixels(b, pixel_type_to[i]);
 
         if (memcmp(a, b, w * h * 1) != 0)
@@ -82,7 +82,7 @@ static int test_mat_pixel_rgb(int w, int h)
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels(a, pixel_type_from[i], w, h);
-        ncnn::Mat b(w, h, 3u, 3);
+        ncnn::Mat b(w, h, (size_t)3u, 3);
         m.to_pixels(b, pixel_type_to[i]);
 
         if (memcmp(a, b, w * h * 3) != 0)
@@ -106,7 +106,7 @@ static int test_mat_pixel_bgr(int w, int h)
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels(a, pixel_type_from[i], w, h);
-        ncnn::Mat b(w, h, 3u, 3);
+        ncnn::Mat b(w, h, (size_t)3u, 3);
         m.to_pixels(b, pixel_type_to[i]);
 
         if (memcmp(a, b, w * h * 3) != 0)
@@ -129,7 +129,7 @@ static int test_mat_pixel_rgba(int w, int h)
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels(a, pixel_type_from[i], w, h);
-        ncnn::Mat b(w, h, 4u, 4);
+        ncnn::Mat b(w, h, (size_t)4u, 4);
         m.to_pixels(b, pixel_type_to[i]);
 
         if (memcmp(a, b, w * h * 4) != 0)
@@ -152,7 +152,7 @@ static int test_mat_pixel_bgra(int w, int h)
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels(a, pixel_type_from[i], w, h);
-        ncnn::Mat b(w, h, 4u, 4);
+        ncnn::Mat b(w, h, (size_t)4u, 4);
         m.to_pixels(b, pixel_type_to[i]);
 
         if (memcmp(a, b, w * h * 4) != 0)
@@ -179,7 +179,7 @@ static int test_mat_pixel_roi_gray(int w, int h, int roix, int roiy, int roiw, i
     for (int i = 0; i < 1; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels_roi(a, pixel_type_from[i], w, h, roix, roiy, roiw, roih);
-        ncnn::Mat b(roiw, roih, 1u, 1);
+        ncnn::Mat b(roiw, roih, (size_t)1u, 1);
         m.to_pixels(b, pixel_type_to[i]);
 
         ncnn::Mat b2;
@@ -211,7 +211,7 @@ static int test_mat_pixel_roi_rgb(int w, int h, int roix, int roiy, int roiw, in
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels_roi(a, pixel_type_from[i], w, h, roix, roiy, roiw, roih);
-        ncnn::Mat b(roiw, roih, 3u, 3);
+        ncnn::Mat b(roiw, roih, (size_t)3u, 3);
         m.to_pixels(b, pixel_type_to[i]);
 
         ncnn::Mat b2;
@@ -243,7 +243,7 @@ static int test_mat_pixel_roi_bgr(int w, int h, int roix, int roiy, int roiw, in
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels_roi(a, pixel_type_from[i], w, h, roix, roiy, roiw, roih);
-        ncnn::Mat b(roiw, roih, 3u, 3);
+        ncnn::Mat b(roiw, roih, (size_t)3u, 3);
         m.to_pixels(b, pixel_type_to[i]);
 
         ncnn::Mat b2;
@@ -274,7 +274,7 @@ static int test_mat_pixel_roi_rgba(int w, int h, int roix, int roiy, int roiw, i
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels_roi(a, pixel_type_from[i], w, h, roix, roiy, roiw, roih);
-        ncnn::Mat b(roiw, roih, 4u, 4);
+        ncnn::Mat b(roiw, roih, (size_t)4u, 4);
         m.to_pixels(b, pixel_type_to[i]);
 
         ncnn::Mat b2;
@@ -305,7 +305,7 @@ static int test_mat_pixel_roi_bgra(int w, int h, int roix, int roiy, int roiw, i
     for (int i = 0; i < 2; i++)
     {
         ncnn::Mat m = ncnn::Mat::from_pixels_roi(a, pixel_type_from[i], w, h, roix, roiy, roiw, roih);
-        ncnn::Mat b(roiw, roih, 4u, 4);
+        ncnn::Mat b(roiw, roih, (size_t)4u, 4);
         m.to_pixels(b, pixel_type_to[i]);
 
         ncnn::Mat b2;
@@ -340,10 +340,10 @@ static int test_mat_pixel_yuv420sp2rgb(int w, int h)
         p += 2;
     }
 
-    ncnn::Mat rgb(w, h, 3u, 3);
+    ncnn::Mat rgb(w, h, (size_t)3u, 3);
     yuv420sp2rgb(nv21, w, h, rgb);
 
-    ncnn::Mat rgb2(w, h, 3u, 3);
+    ncnn::Mat rgb2(w, h, (size_t)3u, 3);
     yuv420sp2rgb_nv12(nv12, w, h, rgb2);
 
     if (memcmp(rgb, rgb2, w * h * 3) != 0)
